requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
<<<<<<< HEAD
Django==3.1.5
Flask==1.1.2
=======
Django==3.1.6
Flask==1.1.2
Flask-Cors==3.0.10
>>>>>>> 45df89bd5fa9c3ad9531332569210ff7d8d5ca79
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gTTS==2.2.1
<<<<<<< HEAD
=======
gunicorn==20.0.4
>>>>>>> 45df89bd5fa9c3ad9531332569210ff7d8d5ca79
h5py==2.10.0
idna==2.10
image==1.5.33
itsdangerous==1.1.0
<<<<<<< HEAD
Jinja2==2.11.2
=======
Jinja2==2.11.3
>>>>>>> 45df89bd5fa9c3ad9531332569210ff7d8d5ca79
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
<<<<<<< HEAD
pytz==2020.5
=======
pytz==2021.1
>>>>>>> 45df89bd5fa9c3ad9531332569210ff7d8d5ca79
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
<<<<<<< HEAD
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
gunicorn== 19.9.0
=======
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
>>>>>>> 45df89bd5fa9c3ad9531332569210ff7d8d5ca79
